requirements.txt
astroid==2.5.3
autopep8==1.5.7
cachetools==4.2.2
certifi==2021.5.30
chardet==4.0.0
click==8.0.1
dnspython==2.1.0
Flask==2.0.1
Flask-PyMongo==2.3.0
google-auth==1.32.1
google-auth-oauthlib==0.4.4
idna==2.10
isort==5.8.0
itsdangerous==2.0.1
Jinja2==3.0.1
lazy-object-proxy==1.6.0
MarkupSafe==2.0.1
mccabe==0.6.1
oauthlib==3.1.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.7.0
pylint==2.7.4
pymongo==3.11.4
requests==2.25.1
requests-oauthlib==1.3.0
rope==0.19.0
rsa==4.7.2
six==1.16.0
toml==0.10.2
urllib3==1.26.6
Werkzeug==2.0.1
wrapt==1.12.1
